backend/requirements.txt
torch>=2.0.0
transformers==4.35.2
accelerate==0.24.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
safetensors==0.4.1
